Extract theme palette logic into src/core/theme.ts
Move theme presets and color derivation functions into a dedicated
module. Typecheck, 167 tests, and build all pass.
